Refactor books endpoints into api/books.py
