+++ server/tabby_server/__main__.py
@@ -1,7 +1,10 @@
 from flask import Flask, request
 from http import HTTPStatus
+
 # from services import resource_format  # Use when actually running the server
 from .services import resource_format  # Use when running pytest
+from .api import books
+
 # import test_routes
 
 # --- IF USING PYTEST VS PYTHON ---
@@ -10,15 +13,19 @@ from .services import resource_format  # Use when running pytest
 
 app = Flask(__name__)
 
+app.register_blueprint(books.subapp, url_prefix="/books")
+
 
 # @app.route('/testpage', methods=['GET'])
 # app.add_url_rule('/testpage', view_func=alternate_page.test_page)
 # @app.route('/test/make_request', methods=['GET'])
-app.add_url_rule('/test/make_request',
-                 view_func=resource_format.google_books_test_call_api)
+app.add_url_rule(
+    "/test/make_request", view_func=resource_format.google_books_test_call_api
+)
 # @app.route('/test/all_books', methods=['GET'])
-app.add_url_rule('/test/all_books',
-                 view_func=resource_format.google_books_test_all_books)
+app.add_url_rule(
+    "/test/all_books", view_func=resource_format.google_books_test_all_books
+)
 
 
 # Only need to swap here. Don't swap it in other files!
@@ -43,48 +50,6 @@ def members():
 @app.route("/api/test", methods=["POST"])
 def test():
     return {}, HTTPStatus.OK
-
-
-@app.route("/books/scan_cover", methods=["POST"])
-def books_scan_cover():
-    """Receives an image and returns a list of possible books that the image
-    could represent.
-
-    Expected fields in JSON:
-    - `"image"`: Base64 data representing the image.
-    """
-
-    if not request.is_json:
-        return {
-            "message": "Content type must be JSON."
-        }, HTTPStatus.BAD_REQUEST
-    data = request.get_json()
-
-    image = data.get("image")
-    if not image:
-        return {
-            "message": "Must specify 'image' as a non-empty string in body."
-        }, HTTPStatus.BAD_REQUEST
-
-    return {"results": []}, HTTPStatus.OK
-
-
-@app.route("/books/search", methods=["GET"])
-def books_search():
-    """Receives a query representing a title and returns a list of possible
-    books that could match.
-
-    Expected query parameters:
-    - `"title"`: Title query from user.
-    """
-
-    title = request.args.get("title")
-    if not title:
-        return {
-            "message": "Must specify 'title' as a non-empty query parameter."
-        }, HTTPStatus.BAD_REQUEST
-
-    return {"results": []}, HTTPStatus.OK  # TODO: remove placeholder
 
 
 if __name__ == "__main__":
